+++ src/main/java/com/Bridgelabz/DigitalSupplyChainTracker/repository/ShipmentRepository.java
@@ -6,5 +6,5 @@ import org.springframework.stereotype.Repository;
 import com.Bridgelabz.DigitalSupplyChainTracker.entity.Shipment;
 
 @Repository
-public interface ShipmentRepository extends JpaRepository<Shipment, Long> {
+public interface ShipmentRepository extends JpaRepository<Shipment, Integer> {
 }
